Add webhook bridge script for poeditor
Poeditor requires one hook per language to be updated. Github limits the
number of hooks to 20. We have 30 languages to updates. Hence, this
scripts listen for incoming webhooks and forwards them to the listed
urls. It doesn't check the content of the hook and relies on the targets
to do that.